// src/app/models/marketplace/nft.model.ts
export class NFTModel {
    nftTokenId: number= 0;
    imagePath:string = '';
    name:string = '';
    description: string='';
    price: number = 0;
    isForSale: boolean = true;
    category: string = '';
    collection: string = '';
    atlasFilePath: string = '';
    network: string = '';
    blockchainId: string = '';
    collectionId: string = '';
    tokenOwner: string = '';
}